service: Split the username input field into username and URL.
The default synergy connector validator has only two fields: username and
password. Given that CardDAV is not a central service with know infrastructure
we also need to specify an URL.

As a fast solution we are using the username input field for username as well as
URL. Combined are both with an @ sign. This commit splits this correctly and
fills in the internals for it.

Long term we want to use a custom validator for this. The docs do not really
describe them so some more investigation is needed before we have this.