fixed customValidatorUI
Issue was that app spawned a new card which did not have parameters.
This is a bit sneaky, because it happens without animation and you only
see it, if you go into card view. The second card captures the focus has
no params and therefore fails during this.params.initialTemplate access.
The original card did work before. Now only the one card in the account
app is existing and account creation works.