disabled postPutRemoteModify
this is because the patch "sync-fix-for-stable-upload" is broken.
It produces an endless loop (which consumes a lot of system ressources
and will drain all battery FAST!), because revs for the local objects
don't get updated correctly. Patch author tried to be too smart... :(
Downside of not having postPutRemoteModify is that etags don't get
stored locally, so the changed objects transmitted to server will be
downloaded again on next downsync, because etag differs.

Open-WebOS-DCO-1.0-Signed-Off-By: Achim Königs <garfonso@tratschtante.de>