added _isOrphanedEntry method
Method checks if DB element belongs to a valid
collection. If not the object probably is orphaned.
In calendar these objects show "undefined" as their
calendar...
This probably happens if calendars are deleted and
recreated.

Orphaned items will get deleted during sync now and
are not used to compare against remote objects, so
if they are still on server, they will get redownloaded,
for example when folder URI changed or event was
moved from one calendar to the other.