path now always is the complete URL for Caldav.js methods
Reworked Caldav.js to now work properly for multiple servers
at the same time.
Changed the API to now include always the full URL in
params.path, so that the http connection can be build on
demand. Also the setHostAndPort method is gone.

This changed folder search a lot.
Also this allowed sendRequest to now retry on a failed
connection and not wait for the timeout.