resolve caldav urls for known servers
known servers can be configured in urlschemes.js.
Basically you define a search string and the corresponding
urls for checkCredentials, calendar-home and addressbook-home.

These will be used during checkCredentials and discovery if
url contians the search string.

The urls can contain %USERNAME% (only once currently) and
a %URL_PREFIX% which will be replaced. URL_PREFIX will be
everything that was before the URL, this is for local
installations of owncloud or egroupdware.