fixed issue with only one activity created
mojosyncframework is stupid here again. Per service it can only create
one sync activity for one kind.
So we changed the sync-callback to be the global sync, i.e. it will sync
contacts and calendar at the same time. For the synconedit to work on
contacts in the future at the same time as calendar, we needed to
overwrite the "complete" method in syncassistant.