added google account
we can add multiple accounts to the package. So the same
js-service is used for all of them, but the limit of
one username per account type is circumvented.
Also configuration for the user can get a lot easier
(google is quite a pain).
I plan to do the same thing for other big stuff like
Facebook, Yahoo and maybe also owncloud. Are there any
objections to this?