fixed issue in on create
if a key already was stored, key manager did recieve an error
and service assistant never continued.
This happens sometimes during create, where onContactsCreate
and onCalendarCreate is called simultaneously.
This leads to an issue where one of the capabilities is never
initialized properly and also the initial sync never happens.