saving log in /media/internal/.org.webosports.cdav.service.log
Sometimes the service seems to hang for days, until it is either
killed or device reboots. During that time all sync attempts
are canceled because "sync already in progress". Also
it consumes quite some memory.. not good.

Usually, when I registered that issue, it was already rotated
out of device logs. So I'll try to get some logs of those hangs.

Log will be cleared every service start.

Signed-off-by: Achim Königs <garfonso@mobo.info>