service: rewrote iCal time handling
Also moved into its own file, this takes a bit of pressure from iCal.js.

Timezone ids are once again respected.
Timestamps are now processed after iCal parsing and  before
iCal generation using TZManager. All timestamps in an event
are considered.

This surfaced a bug in the calendar app:
It can not handle exceptions to events in foreign (i.e. not
UTC and not local) timezones.

Signed-off-by: Achim Königs <garfonso@mobo.info>