fixed: google accounts on webos 2.1.
for some reason 2.1 can not get the username from google.
No clue why. Status is always 0, responseText empty.
So we fill the username with the current timestamp.
The user can rename the account later to something fitting.

But we also needed to fix the base URL to make discovery
work, because resolved urls contain the wrong username now.

Signed-off-by: Achim Königs <garfonso@mobo.info>