iCal: rewrote parseOneLineMethod
the old one had issues handling parameter values with : or ;
and was a bit rough. It already included a strange work
around for : in the value part.

The new method processes the line much more structured and
should now correctly process all allowed lines which the old one
did not.

Signed-off-by: Achim Königs <garfonso@mobo.info>